requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
